Update department.py
writing the program for given, when, then: first we need to set the @given decorator which will display our string with our given context table from our feature. next is the @when which will display a count method from our company_model which we put as pass since right now we do not want to display anything.  we have two @then decorators which will display how many people in the department with two users.  will run a method NameNumber from testutil.py and display the given string in our dictionary.  then assert that count we find and display it with the department.  lastly we will run the method to display one person in another department, assert that to 1 which would be the count for string "one", equal it to the method headcount(department) which displays the current department.